third_party/WebKit/LayoutTests/http/tests/devtools/elements/styles/edit-css-with-source-url-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests file system project mappings.


Running: testEditingRulesInElementsPanelDoesNotAddSourceURLToOriginalFile
Dumping uiSourceCode content:
#inspected {
          color: red;
      }
      
Dumping matched rules:
[expanded] 
element.style { ()

[expanded] 
#inspected { (foo.css:1 -> foo.css:1:13)
    color: red;

[expanded] 
div { (user agent stylesheet)
    display: block;

Editing styles from elements panel:
Styles edited.
Dumping matched rules:
[expanded] 
element.style { ()

[expanded] 
#inspected { (foo.css:1 -> foo.css:1:13)
    color: green;
    : ;

[expanded] 
div { (user agent stylesheet)
    display: block;

Dumping uiSourceCode content:
#inspected {
          color: green;
      }
      
